speech-recognition/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.13.1
idna==3.4
langdetect==1.0.9
llvmlite==0.39.1
more-itertools==9.1.0
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@ad3250a846fe7553a25064a2dc593e492dadf040
packaging==23.0
Pillow==9.4.0
PyAudio==0.2.13
python-osc==1.8.1
PyYAML==6.0
pydub==0.25.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
SpeechRecognition==3.9.0
tokenizers==0.13.2
torch
torchaudio
torchvision
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.15
wincertstore==0.2
yarl==1.8.2
